requirements.txt
wheel
streamlit
llama-index
boto3
python-dotenv
bs4